Finalização do CRU
CRUD com todas as ações, mesmo que seja tudo na controller...

+++ src/main/java/br/edu/ifrs/restinga/dev1/alonso/demo/Repository/LivroRepository.java
@@ -3,9 +3,8 @@ package br.edu.ifrs.restinga.dev1.alonso.demo.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import br.edu.ifrs.restinga.dev1.alonso.demo.Model.Livro;
+import br.edu.ifrs.restinga.dev1.alonso.demo.Model.Objeto.Livro;
 
 @Repository
-public interface LivroRepository extends JpaRepository<Livro, Long>{
-	
+public interface LivroRepository extends JpaRepository<Livro, Integer>{	
 }
